Extracted biomedsheet parser methods to it own class.
